Support for remaining SOS variables
Reworking of VariableParams to use an enumeration rather than static fields - this  facilitates easier lookups and quicker addition of new variables.  Also added the remaining "CF Compliant" SOS variables.  Non-cf compliant variables are still included, but are not given a standard name.
Improved the determination of header fields in SosAgent - now properly locates the "depth (m)" header (if present) and adjusts the start of the data columns, which ALWAYS come after "depth (m)" or at the 5th column.
Added logic in NcFactory to deal with a "null" standard_name (by simply excluding the attribute from the variable)
The refactoring of VariableParams resulted in a couple of small name/form changes in a few other classes.